JIRA: GP-25818 Change ValidBlob arg to unsigned to eliminate warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@84877 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objects/genomecoll/cached_assembly.hpp
@@ -45,7 +45,7 @@ public:
     const string& Blob();
 
 public:
-    static bool ValidBlob(int blobSize);
+    static bool ValidBlob(size_t blobSize);
     static CCompressStream::EMethod Compression(const string& blob);
 
 private:
